+++ changelog.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added 
 - Added Film Grain intensity slider.
 - Added Support for 1.7.10 Streams mod.
+- Added nether biome fog color support.
 
 ### Fixed